+++ CHANGELOG.md
@@ -88,6 +88,7 @@ IMPROVEMENTS:
   - `MAX_MQ_MESSAGE_LENGTH`: Maximum MQ message size (in bytes) (after compression if compressed). This value must be less than `MAX_MQ_RECV_MESSAGE_SIZE` in MQ service configuration on receiving side. Default to `3251200` (3.1MB).
   - `AS_DATA_MAX_UNCOMPRESSED_LENGTH`: Maximum AS response data size (in bytes) when uncompressed (before compression / after decompression). This configuration should be the same on both sending and receiving parties. Default to `10485760` (10MB).
   - `AS_DATA_MAX_LENGTH`: Maximum AS response data size (in bytes) (after compression if compressed). This value must be less than `MAX_MQ_MESSAGE_LENGTH` to leave sufficient headroom for the rest of the message payload. Default to `3145728` (3MB).
+  - `HTTP_REQUEST_BODY_SIZE_LIMIT`: Maximum HTTP request body size (in bytes). Default to `26214400` (25MB).
 - [MQ Service] Add new environment variable options
   - `MAX_MQ_RECV_MESSAGE_SIZE`: Maximum inbound message size (in bytes) on receiving socket. Default to `3300000`.
 - Update dependencies.